update env plugin
